+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransactionType.java
@@ -22,6 +22,9 @@ public enum TransactionType {
     REGISTER_DATA((byte) 21),
     TRANSFER_WITH_MEMO((byte) 22);
 
+    /**
+     * No of Bytes used for Serializing {@link TransactionType}.
+     */
     public static final int BYTES = 1;
 
     @Getter