+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/smartcontracts/WasmModule.java
@@ -1,70 +1,106 @@
 package com.concordium.sdk.transactions.smartcontracts;
 
 import com.concordium.sdk.crypto.SHA256;
+import com.concordium.sdk.responses.modulelist.ModuleRef;
 import com.concordium.sdk.transactions.TransactionType;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.val;
-import org.apache.commons.codec.binary.Hex;
 
 import java.nio.ByteBuffer;
 import java.util.Arrays;
-import java.util.Objects;
 
+/**
+ * A compiled Smart Contract Module in WASM with source and version.
+ */
 @Getter
 public class WasmModule {
 
-    private final WasmModuleVersion version;
-    private final WasmModuleSource moduleSource;
+    private static final long MAX_SIZE_V0 = 65536; // 64 kb
+    private static final long MAX_SIZE_V1 = 8 * 65536; // 512 kb
 
-    @Builder
-    public WasmModule(WasmModuleVersion version, WasmModuleSource moduleSource) {
-        if (Objects.isNull(moduleSource)) {
-            throw new IllegalArgumentException("WasmModuleSource cannot be null");
-        }
+    /**
+     * No of bytes reserved to keep Module Version Information.
+     */
+    static final int VERSION_BYTES = 8;
+
+
+    /**
+     * Module version.
+     */
+    private final WasmModuleVersion version;
+
+    /**
+     * Module Source
+     */
+    private final WasmModuleSource source;
+
+    WasmModule(final WasmModuleSource source, final WasmModuleVersion version) {
+        this.source = source;
         this.version = version;
-        this.moduleSource = moduleSource;
     }
 
-    public static WasmModule from(byte[] fileBuffer, int version) {
-        WasmModuleSource moduleSource;
-        WasmModuleVersion moduleVersion;
-        fileBuffer = Arrays.copyOfRange(fileBuffer, 8, fileBuffer.length);
+    /**
+     * Create {@link WasmModule} from compiled module WASM file bytes.
+     *
+     * @param bytes   Complied WASM module file bytes.
+     * @param version WASM module version ({@link WasmModuleVersion}). If the version is prefixed use {@link WasmModule#from(byte[])}
+     * @return Parsed {@link WasmModule}
+     */
+    public static WasmModule from(final byte[] bytes, final WasmModuleVersion version) {
+        switch (version) {
+            case V0:
+                if (bytes.length > MAX_SIZE_V0) {
+                    throw new IllegalArgumentException("Wasm V0 modules may not exceed " + MAX_SIZE_V0 + " bytes.");
+                }
+                break;
+            case V1:
+                if (bytes.length > MAX_SIZE_V1) {
+                    throw new IllegalArgumentException("Wasm V1 modules may not exceed " + MAX_SIZE_V1 + " bytes.");
+                }
+                break;
+        }
 
-        if (version == 1) {
-            moduleSource = WasmModuleSourceV1.from(fileBuffer);
-            moduleVersion = WasmModuleVersion.V1;
-        }
-        else {
-            moduleSource = WasmModuleSourceV0.from(fileBuffer);
-            moduleVersion = WasmModuleVersion.V0;
-        }
-        return new WasmModule(moduleVersion, moduleSource);
+        return new WasmModule(WasmModuleSource.from(bytes), version);
+    }
+
+    /**
+     * Create {@link WasmModule} from compiled module WASM file bytes. Passed module bytes should have version prefixed.
+     *
+     * @param bytes Complied WASM module file bytes.
+     * @return Parsed {@link WasmModule}
+     */
+    public static WasmModule from(final byte[] bytes) {
+        val versionBytes = ByteBuffer.wrap(Arrays.copyOfRange(bytes, 0, VERSION_BYTES));
+        val version = WasmModuleVersion.from(versionBytes);
+        val moduleBytes = Arrays.copyOfRange(bytes, VERSION_BYTES, bytes.length);
+
+        return from(moduleBytes, version);
     }
 
     /**
      * Get the identifier of the WasmModule.
      * The identifier is a SHA256 hash of the raw module bytes.
+     *
      * @return the identifier of the Module.
      */
-    public String getIdentifier() {
-        return Hex.encodeHexString(SHA256.hash(this.moduleSource.getBytes()));
+    public ModuleRef getIdentifier() {
+        return ModuleRef.from(SHA256.hash(this.getBytes()));
     }
 
     /**
      * Get the raw serialized bytes of the concrete {@link WasmModule}.
-     * This can currently either be {@link WasmModuleSourceV0} or {@link WasmModuleSourceV1}
      *
      * @return bytes
      */
-
     public byte[] getBytes() {
-        val moduleSourceBytes = moduleSource.getBytes();
-        val versionBytes = version.getBytes();
-        val buffer = ByteBuffer.allocate(TransactionType.BYTES + moduleSourceBytes.length + versionBytes.length);
+        val moduleSourceBytes = source.getBytes();
+        val buffer = ByteBuffer.allocate(TransactionType.BYTES +
+                moduleSourceBytes.length +
+                WasmModuleVersion.BYTES);
         buffer.put(TransactionType.DEPLOY_MODULE.getValue());
-        buffer.put(versionBytes);
+        buffer.put(version.getBytes());
         buffer.put(moduleSourceBytes);
+
         return buffer.array();
     }
 }